MAINT: More refactoring
