+++ openff/interchange/_tests/interoperability_tests/components/test_packmol.py
@@ -1,11 +1,11 @@
 import pytest
-from openff.toolkit import Quantity
-
-from openff.interchange._tests import MoleculeWithConformer
-from openff.interchange.components._packmol import (
+from openff.packmol import (
     _find_packmol,
     solvate_topology_nonwater,
 )
+from openff.toolkit import Quantity
+
+from openff.interchange._tests import MoleculeWithConformer
 from openff.interchange.drivers import get_openmm_energies
