+++ openff/interchange/components/_packmol.py
@@ -6,11 +6,6 @@ from openff.packmol import (
     RHOMBIC_DODECAHEDRON,
     RHOMBIC_DODECAHEDRON_XYHEX,
     UNIT_CUBE,
-    pack_box,
-    solvate_topology,
-    solvate_topology_nonwater,
-)
-from openff.packmol._packmol import (
     _box_from_density,
     _box_vectors_are_in_reduced_form,
     _build_input_file,
@@ -31,6 +26,9 @@ from openff.packmol._packmol import (
     _validate_inputs,
     _wrap_into,
     _wrap_into_brick,
+    pack_box,
+    solvate_topology,
+    solvate_topology_nonwater,
 )
 
 __all__ = (
